perf(fetch): store response body as Bytes; fix arrayBuffer binary corruption
Store FetchResponse.body as bytes::Bytes instead of Vec<u8>. The body is
filled once from reqwest's response.bytes() and read by several accessors
(text/json/arrayBuffer/bytes/blob/formData), each of which cloned the whole
payload to release the registry lock before allocating into the arena. With
Bytes that clone is an O(1) refcount bump, so res.json()/res.text() no longer
copy the body on the hot path; the text/json decode is folded into a single
arena write (from_utf8_lossy borrows on valid UTF-8).

Fix the arrayBuffer non-UTF-8 corruption: js_response_array_buffer and
js_blob_array_buffer round-tripped the body through from_utf8_unchecked into
a String and resolved a STRING_TAG value. That is UB on non-UTF-8 bytes (a
fetched PNG/protobuf), and the JS new Uint8Array(value) dispatch keys on the
Buffer POINTER_TAG, so a string-tagged value was read as an empty buffer.
Both paths now resolve a real Buffer via alloc_buffer, byte-exact, matching
the existing bytes() accessor.

Remove three debug eprintln! calls, including one in js_fetch_response_status
that ran on every status read and collected the response-id map under the lock.

Tests: unit regression for the byte-exact buffer resolve (fail-before proven)
plus text/json output preservation, and an e2e shell test exercising the full
Blob.arrayBuffer / Response.text / Response.json / Response.arrayBuffer chain.
