+++ crates/perry-ext-fetch/Cargo.toml
@@ -11,6 +11,7 @@ crate-type = ["staticlib", "rlib"]
 [dependencies]
 perry-ffi.workspace = true
 reqwest = { version = "0.12", features = ["json", "rustls-tls", "http2"], default-features = false }
+bytes = "1"
 tokio = { workspace = true }
 serde_json = "1"
 lazy_static = "1.5"
